requirements.txt
fastapi
pillow
customtkinter
tensorflow==2.20.0
numpy
pandas
matplotlib
opencv-python
scikit-learn
uvicorn
